Fixed duplicate protocol buffer source in default repo.
All proto files present in src/ and third_party/runtime are now staged in a single folder. This is used as the base folder for the proto compiler.

This alleviates the problem that the federation proto that is maintained in libraries was manually copied over.

Reviewed by: Alex
Rietveld Issue: 186001

Fixes Issue 111
Update Issue 111
Labels: Type-Enhancement